+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Mobilisten Plugin Changelog
 
+### 1.0.2 - 21 Feb 2022
+
+- Updated Mobilisten SDK for iOS to version 4.2.6. [Learn More about the update](https://github.com/zoho/SalesIQ-Mobilisten-iOS/releases/tag/v4.2.6).
+- Updated Mobilisten SDK for Android to version 4.2.3. [Learn More about the update](https://github.com/zoho/salesiq-mobilisten-android-sample/releases/tag/4.2.3).
+- Added **.chatUnreadCount** API to fetch the current unread count of the chats. [Learn more](https://www.zoho.com/salesiq/help/developer-guides/flutter-sdk-chat-unread-count.html).
+- Added **chatUnreadCountChanged** event in chatEventChannel to track the change in the unreadCount. [Learn more](https://www.zoho.com/salesiq/help/developer-guides/flutter-sdk-event-handler-chatEventChannel.html).
+
 ### 1.0.1 - 08 Dec 2021
 
 - Updated Mobilisten SDK for iOS to version 4.2.3. [Learn More about the update](https://github.com/zoho/SalesIQ-Mobilisten-iOS/releases/tag/v4.2.3).